A lot of work done
Added the ability to work with the server directly.
It will now check if it are supposed to start working with the server
right away, by using the last settings.

The buttons will now change function depending on the checkbox is
checked or not, being if it is off, it will work locally, on it will
work with the server.

The shoppingList were modified to work with the gui, and store wether
the checkboxed is stored or not.